# Speed up decoding of large enums