Generic `to` is quite slow on datatypes with many constructors, because it runs `O(log^2(n))` `instanceof` operations, where `n` is the number of constructors.

This change moves this cost from the decoding loop to initialization time. Now we store values of the target type in the lookup table instead of Generic representations. This makes initialization slower, but if there are many enums values to decode, we save time there. 

Before:

```
Enum3 Unscramble                        : 0.000027ms/op
Enum10 Unscramble                       : 0.000169ms/op
Enum30 Unscramble                       : 0.002258ms/op
```

After:

```
Enum3 Unscramble                        : 0.000016ms/op
Enum10 Unscramble                       : 0.000047ms/op
Enum30 Unscramble                       : 0.000050ms/op
```

As seen in the above benchmark results, after this change enum decoding is almost independent of enum size. Which makes sense, because now it's only a hashtable lookup, as it should be.